+++ pkg/repository/sqlcv1/workflows.sql.go
@@ -926,7 +926,8 @@ INSERT INTO "WorkflowVersion" (
     "inputJsonSchema",
     "idempotencyKeyExpression",
     "idempotencyKeyTtlMs",
-    "displayName"
+    "displayName",
+    "idempotencyMethod"
 ) VALUES (
     $1::uuid,
     coalesce($2::timestamp, CURRENT_TIMESTAMP),
@@ -944,26 +945,28 @@ INSERT INTO "WorkflowVersion" (
     $12::jsonb,
     $13::text,
     $14::bigint,
-    $15::text
-) RETURNING id, "createdAt", "updatedAt", "deletedAt", version, "order", "workflowId", checksum, "scheduleTimeout", "onFailureJobId", sticky, kind, "defaultPriority", "createWorkflowVersionOpts", "inputJsonSchema", "idempotencyKeyExpression", "idempotencyKeyTtlMs", "displayName"
+    $15::text,
+    $16::idempotency_method
+) RETURNING id, "createdAt", "updatedAt", "deletedAt", version, "order", "workflowId", checksum, "scheduleTimeout", "onFailureJobId", sticky, kind, "defaultPriority", "createWorkflowVersionOpts", "inputJsonSchema", "idempotencyKeyExpression", "idempotencyKeyTtlMs", "displayName", "idempotencyMethod"
 `
 
 type CreateWorkflowVersionParams struct {
-	ID                        uuid.UUID          `json:"id"`
-	CreatedAt                 pgtype.Timestamp   `json:"createdAt"`
-	UpdatedAt                 pgtype.Timestamp   `json:"updatedAt"`
-	Deletedat                 pgtype.Timestamp   `json:"deletedat"`
-	Checksum                  string             `json:"checksum"`
-	Version                   pgtype.Text        `json:"version"`
-	Workflowid                uuid.UUID          `json:"workflowid"`
-	Sticky                    NullStickyStrategy `json:"sticky"`
-	Kind                      NullWorkflowKind   `json:"kind"`
-	DefaultPriority           pgtype.Int4        `json:"defaultPriority"`
-	CreateWorkflowVersionOpts []byte             `json:"createWorkflowVersionOpts"`
-	InputJsonSchema           []byte             `json:"inputJsonSchema"`
-	IdempotencyKeyExpression  pgtype.Text        `json:"idempotencyKeyExpression"`
-	IdempotencyKeyTtlMs       pgtype.Int8        `json:"idempotencyKeyTtlMs"`
-	DisplayName               pgtype.Text        `json:"displayName"`
+	ID                        uuid.UUID             `json:"id"`
+	CreatedAt                 pgtype.Timestamp      `json:"createdAt"`
+	UpdatedAt                 pgtype.Timestamp      `json:"updatedAt"`
+	Deletedat                 pgtype.Timestamp      `json:"deletedat"`
+	Checksum                  string                `json:"checksum"`
+	Version                   pgtype.Text           `json:"version"`
+	Workflowid                uuid.UUID             `json:"workflowid"`
+	Sticky                    NullStickyStrategy    `json:"sticky"`
+	Kind                      NullWorkflowKind      `json:"kind"`
+	DefaultPriority           pgtype.Int4           `json:"defaultPriority"`
+	CreateWorkflowVersionOpts []byte                `json:"createWorkflowVersionOpts"`
+	InputJsonSchema           []byte                `json:"inputJsonSchema"`
+	IdempotencyKeyExpression  pgtype.Text           `json:"idempotencyKeyExpression"`
+	IdempotencyKeyTtlMs       pgtype.Int8           `json:"idempotencyKeyTtlMs"`
+	DisplayName               pgtype.Text           `json:"displayName"`
+	IdempotencyMethod         NullIdempotencyMethod `json:"idempotencyMethod"`
 }
 
 func (q *Queries) CreateWorkflowVersion(ctx context.Context, db DBTX, arg CreateWorkflowVersionParams) (*WorkflowVersion, error) {
@@ -983,6 +986,7 @@ func (q *Queries) CreateWorkflowVersion(ctx context.Context, db DBTX, arg Create
 		arg.IdempotencyKeyExpression,
 		arg.IdempotencyKeyTtlMs,
 		arg.DisplayName,
+		arg.IdempotencyMethod,
 	)
 	var i WorkflowVersion
 	err := row.Scan(
@@ -1004,6 +1008,7 @@ func (q *Queries) CreateWorkflowVersion(ctx context.Context, db DBTX, arg Create
 		&i.IdempotencyKeyExpression,
 		&i.IdempotencyKeyTtlMs,
 		&i.DisplayName,
+		&i.IdempotencyMethod,
 	)
 	return &i, err
 }
@@ -1258,7 +1263,7 @@ func (q *Queries) GetWorkflowShape(ctx context.Context, db DBTX, workflowversion
 
 const getWorkflowVersionById = `-- name: GetWorkflowVersionById :one
 SELECT
-    wv.id, wv."createdAt", wv."updatedAt", wv."deletedAt", wv.version, wv."order", wv."workflowId", wv.checksum, wv."scheduleTimeout", wv."onFailureJobId", wv.sticky, wv.kind, wv."defaultPriority", wv."createWorkflowVersionOpts", wv."inputJsonSchema", wv."idempotencyKeyExpression", wv."idempotencyKeyTtlMs", wv."displayName",
+    wv.id, wv."createdAt", wv."updatedAt", wv."deletedAt", wv.version, wv."order", wv."workflowId", wv.checksum, wv."scheduleTimeout", wv."onFailureJobId", wv.sticky, wv.kind, wv."defaultPriority", wv."createWorkflowVersionOpts", wv."inputJsonSchema", wv."idempotencyKeyExpression", wv."idempotencyKeyTtlMs", wv."displayName", wv."idempotencyMethod",
     w.id, w."createdAt", w."updatedAt", w."deletedAt", w."tenantId", w.name, w.description, w."isPaused"
 FROM
     "WorkflowVersion" as wv
@@ -1296,6 +1301,7 @@ func (q *Queries) GetWorkflowVersionById(ctx context.Context, db DBTX, id uuid.U
 		&i.WorkflowVersion.IdempotencyKeyExpression,
 		&i.WorkflowVersion.IdempotencyKeyTtlMs,
 		&i.WorkflowVersion.DisplayName,
+		&i.WorkflowVersion.IdempotencyMethod,
 		&i.Workflow.ID,
 		&i.Workflow.CreatedAt,
 		&i.Workflow.UpdatedAt,
@@ -1384,7 +1390,7 @@ func (q *Queries) GetWorkflowVersionEventTriggerRefs(ctx context.Context, db DBT
 
 const getWorkflowVersionForEngine = `-- name: GetWorkflowVersionForEngine :many
 SELECT
-    workflowversions.id, workflowversions."createdAt", workflowversions."updatedAt", workflowversions."deletedAt", workflowversions.version, workflowversions."order", workflowversions."workflowId", workflowversions.checksum, workflowversions."scheduleTimeout", workflowversions."onFailureJobId", workflowversions.sticky, workflowversions.kind, workflowversions."defaultPriority", workflowversions."createWorkflowVersionOpts", workflowversions."inputJsonSchema", workflowversions."idempotencyKeyExpression", workflowversions."idempotencyKeyTtlMs", workflowversions."displayName",
+    workflowversions.id, workflowversions."createdAt", workflowversions."updatedAt", workflowversions."deletedAt", workflowversions.version, workflowversions."order", workflowversions."workflowId", workflowversions.checksum, workflowversions."scheduleTimeout", workflowversions."onFailureJobId", workflowversions.sticky, workflowversions.kind, workflowversions."defaultPriority", workflowversions."createWorkflowVersionOpts", workflowversions."inputJsonSchema", workflowversions."idempotencyKeyExpression", workflowversions."idempotencyKeyTtlMs", workflowversions."displayName", workflowversions."idempotencyMethod",
     w."name" as "workflowName",
     wc."limitStrategy" as "concurrencyLimitStrategy",
     wc."maxRuns" as "concurrencyMaxRuns",
@@ -1445,6 +1451,7 @@ func (q *Queries) GetWorkflowVersionForEngine(ctx context.Context, db DBTX, arg 
 			&i.WorkflowVersion.IdempotencyKeyExpression,
 			&i.WorkflowVersion.IdempotencyKeyTtlMs,
 			&i.WorkflowVersion.DisplayName,
+			&i.WorkflowVersion.IdempotencyMethod,
 			&i.WorkflowName,
 			&i.ConcurrencyLimitStrategy,
 			&i.ConcurrencyMaxRuns,
@@ -1511,7 +1518,7 @@ const linkOnFailureJob = `-- name: LinkOnFailureJob :one
 UPDATE "WorkflowVersion"
 SET "onFailureJobId" = $1::uuid
 WHERE "id" = $2::uuid
-RETURNING id, "createdAt", "updatedAt", "deletedAt", version, "order", "workflowId", checksum, "scheduleTimeout", "onFailureJobId", sticky, kind, "defaultPriority", "createWorkflowVersionOpts", "inputJsonSchema", "idempotencyKeyExpression", "idempotencyKeyTtlMs", "displayName"
+RETURNING id, "createdAt", "updatedAt", "deletedAt", version, "order", "workflowId", checksum, "scheduleTimeout", "onFailureJobId", sticky, kind, "defaultPriority", "createWorkflowVersionOpts", "inputJsonSchema", "idempotencyKeyExpression", "idempotencyKeyTtlMs", "displayName", "idempotencyMethod"
 `
 
 type LinkOnFailureJobParams struct {
@@ -1541,6 +1548,7 @@ func (q *Queries) LinkOnFailureJob(ctx context.Context, db DBTX, arg LinkOnFailu
 		&i.IdempotencyKeyExpression,
 		&i.IdempotencyKeyTtlMs,
 		&i.DisplayName,
+		&i.IdempotencyMethod,
 	)
 	return &i, err
 }