+++ sdks/python/hatchet_sdk/runnables/types.py
@@ -18,7 +18,10 @@ from hatchet_sdk.contracts.v1.workflows_pb2 import DefaultFilter as DefaultFilte
 from hatchet_sdk.types.concurrency import (
     ConcurrencyExpression,
 )
-from hatchet_sdk.types.idempotency import TTLBasedIdempotencyConfig
+from hatchet_sdk.types.idempotency import (
+    StatusBasedIdempotencyConfig,
+    TTLBasedIdempotencyConfig,
+)
 from hatchet_sdk.types.priority import Priority
 from hatchet_sdk.types.sticky import StickyStrategy
 from hatchet_sdk.utils.timedelta_to_expression import Duration
@@ -107,7 +110,7 @@ class WorkflowConfig(BaseModel):
     display_name: str | None = None
     input_validator: TypeAdapter[TaskPayloadForInternalUse]
     default_priority: int | Priority | None = None
-    idempotency: TTLBasedIdempotencyConfig | None = None
+    idempotency: TTLBasedIdempotencyConfig | StatusBasedIdempotencyConfig | None = None
 
     task_defaults: TaskDefaults = TaskDefaults()
     default_filters: list[DefaultFilter] = Field(default_factory=list)